feat(terrain): expose northness/eastness zone axes; harden stats numerics
The aspect components were generated but unqueryable (no zoning scheme).
Split the 2-band aspect_components.tif into single-band northness.tif /
eastness.tif -- the tiled reader reads band 0 only, so per-band keys would
have contorted the ZoneLayer model -- each with a BandedZoning over a
[-100, 100] percent domain (default 50-pct bands; band_step_pct to
adjust). Terrain format version 2 -> 3; provenance forces the rebuild.
The COSSIN_SLOPE_WEIGHTED module constant becomes a real
GenerationOptions.cossin_slope_weighted option instead of a dead knob.

Stats hardening: DatasetVariable now rejects NaN nodata at construction
(the != fill filter cannot exclude NaN, which would poison reductions),
and the zonal-stats results array is float64 -- reductions already
computed in float64 and were truncated to ~7 significant digits on
storage. Tests pinning float32-masked values move to pytest.approx.

+++ src/snowtool/snowdb/zonal_stats.py
@@ -133,6 +133,11 @@ class ZonalStats:
         # preserves that order rather than re-sorting.
         self._cells_index = {cell: idx for idx, cell in enumerate(zone_cells)}
         self._dates_index = {dt: idx for idx, dt in enumerate(sorted(dates))}
+        # float64, not float32: the per-cell reduction runs in float64
+        # (_ZoneIndex.reduce), and area/total stats reach ~1e9-scale values that
+        # float32 truncates to ~7 significant digits. The array is tiny
+        # (cells x dates x stats) and JSON output is float64 anyway, so store the
+        # full precision rather than round-tripping through float32.
         self._array = numpy.full(
             (
                 len(self._dates_index),
@@ -140,7 +145,7 @@ class ZonalStats:
                 len(self._variables_index) + 1,
             ),
             -numpy.inf,
-            dtype=numpy.float32,
+            dtype=numpy.float64,
         )
 
         self.add_results(*results)
